# Conflicts:
#	build_scripts/build_usd.py
#	pxr/base/lib/trace/jsonSerialization.cpp
#	pxr/imaging/lib/garch/CMakeLists.txt
#	pxr/imaging/lib/garch/textureRegistry.cpp
#	pxr/imaging/lib/glf/plugInfo.json
#	pxr/imaging/lib/glf/ptexTexture.h
#	pxr/imaging/lib/hdSt/CMakeLists.txt
#	pxr/imaging/lib/hdSt/GL/glUtils.cpp
#	pxr/imaging/lib/hdSt/codeGen.cpp
#	pxr/imaging/lib/hdSt/extCompGpuComputation.h
#	pxr/imaging/lib/hdSt/geometricShader.cpp
#	pxr/imaging/lib/hdSt/glslProgram.cpp
#	pxr/imaging/lib/hdSt/immediateDrawBatch.cpp
#	pxr/imaging/lib/hdSt/indirectDrawBatch.cpp
#	pxr/imaging/lib/hdSt/package.cpp
#	pxr/imaging/lib/hdSt/renderPassState.cpp
#	pxr/imaging/lib/hdSt/shaders/renderPass.glslfx
#	pxr/imaging/lib/hdSt/shaders/terminals.glslfx
#	pxr/imaging/lib/hdSt/unitTestHelper.cpp
#	pxr/imaging/lib/hdx/drawTargetRenderPass.cpp
#	pxr/imaging/lib/hdx/drawTargetTask.cpp
#	pxr/imaging/lib/hdx/intersector.cpp
#	pxr/imaging/lib/hdx/intersector.h
#	pxr/imaging/lib/hdx/shaders/renderPass.glslfx
#	pxr/imaging/lib/hdx/shadowTask.cpp
#	pxr/imaging/lib/hdx/simpleLightTask.cpp
#	pxr/imaging/lib/hdx/testenv/testHdxIdRender.cpp
#	pxr/usd/lib/usd/stage.h
#	pxr/usd/lib/usdSkel/blendShape.cpp
#	pxr/usdImaging/lib/usdImagingGL/engine.cpp
#	pxr/usdImaging/lib/usdImagingGL/engine.h
#	pxr/usdImaging/lib/usdImagingGL/wrapEngine.cpp
#	pxr/usdImaging/lib/usdviewq/appController.py
#	third_party/maya/lib/pxrUsdMayaGL/batchRenderer.cpp
#	third_party/maya/lib/pxrUsdMayaGL/batchRenderer.h
#	third_party/maya/lib/pxrUsdMayaGL/proxyShapeDelegate.cpp

+++ pxr/imaging/lib/hdSt/GL/glUtils.cpp
@@ -62,7 +62,7 @@ _CreateVtArray(int numElements, int arraySize, int stride,
 
     TF_VERIFY(data.size() == stride*(numElements-1) + arraySize*sizeof(T));
 
-    if (stride == arraySize*sizeof(T)) {
+    if (stride == static_cast<int>(arraySize*sizeof(T))) {
         memcpy(dst, src, numElements*arraySize*sizeof(T));
     } else {
         // deinterleaving
